+++ packages/server/admin/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./services/profileService";
+export * from "./services/membershipService";
 export * from "./services/eventService";
 export * from "./services/applicationService";
 export * from "./services/webhookService";
