+++ packages/server/admin/src/repositories/profileRepository.ts
@@ -1,11 +1,5 @@
 import { BaseRepository } from "@uwdsc/db/baseRepository";
-import {
-  MarkAsPaidData,
-  Member,
-  MembershipStats,
-  Profile,
-  UserRole,
-} from "@uwdsc/common/types";
+import { Member, Profile, UserRole } from "@uwdsc/common/types";
 
 export class ProfileRepository extends BaseRepository {
   /**
@@ -111,60 +105,6 @@ export class ProfileRepository extends BaseRepository {
       return result[0] ?? null;
     } catch (error: unknown) {
       console.error("Error fetching profile by ID:", error);
-      throw error;
-    }
-  }
-
-  /**
-   * Get membership statistics
-   */
-  async getMembershipStats(): Promise<MembershipStats> {
-    try {
-      const result = await this.sql<MembershipStats[]>`
-        SELECT 
-          COUNT(*) as total_users,
-          COUNT(*) FILTER (WHERE m.profile_id IS NOT NULL) as paid_users,
-          COUNT(*) FILTER (WHERE m.profile_id IS NOT NULL AND p.is_math_soc_member = true) as math_soc_members
-        FROM profiles p
-        LEFT JOIN memberships m ON p.id = m.profile_id
-      `;
-
-      return (
-        result[0] ?? {
-          total_users: 0,
-          paid_users: 0,
-          math_soc_members: 0,
-        }
-      );
-    } catch (error: unknown) {
-      console.error("Error fetching membership stats:", error);
-      throw error;
-    }
-  }
-
-  /**
-   * Mark a member as paid by profile ID
-   * @param profileId - The profile ID (UUID)
-   * @param data - Payment data (method, location, verifier)
-   */
-  async markAsPaid(profileId: string, data: MarkAsPaidData): Promise<boolean> {
-    try {
-      const result = await this.sql`
-        INSERT INTO memberships (profile_id, payment_method, payment_location, verifier_id, verified_at, updated_at)
-        VALUES (${profileId}, ${data.payment_method}::payment_method_enum, ${data.payment_location}, ${data.verifier}::uuid, NOW(), NOW())
-        ON CONFLICT (profile_id) 
-        DO UPDATE SET
-          payment_method = ${data.payment_method}::payment_method_enum,
-          payment_location = ${data.payment_location},
-          verifier_id = ${data.verifier}::uuid,
-          verified_at = NOW(),
-          updated_at = NOW()
-        RETURNING *
-      `;
-
-      return result.length > 0;
-    } catch (error: unknown) {
-      console.error("Error marking member as paid:", error);
       throw error;
     }
   }
