+++ packages/server/admin/src/services/emailService.ts
@@ -1,10 +1,9 @@
 import { render } from "@react-email/render";
 import { createElement } from "react";
 import { Resend } from "resend";
-import { ApiError, type UserRole } from "@uwdsc/common/types";
+import { ApiError } from "@uwdsc/common/types";
 import { CampaignEmailTemplate } from "../email-templates/campaign";
 import { appendMarketingUnsubscribeFooter } from "../utils/marketingEmail";
-import { profileService } from "./profileService";
 
 class EmailService {
   private readonly resend: Resend | null;
@@ -129,21 +128,19 @@ class EmailService {
 
   /**
    * Send a markdown campaign as a Resend **Broadcast** (marketing): adds
-   * recipients to `RESEND_CAMPAIGN_SEGMENT_ID`, sends once, then returns
-   * `recipientEmails` so the caller can run {@link removeRecipientsFromSegment}
+   * `recipientEmails` to `RESEND_CAMPAIGN_SEGMENT_ID`, sends once, then returns
+   * the same list so the caller can run {@link removeRecipientsFromSegment}
    * after a delay (broadcast audience may be resolved asynchronously on Resend’s side).
    *
    * Create a dedicated empty segment in the Resend dashboard and set
    * `RESEND_CAMPAIGN_SEGMENT_ID` to its id.
    */
-  async sendCampaignEmail(input: {
-    subject: string;
-    body: string;
-    recipientRoles: UserRole[];
-  }): Promise<{ id?: string; recipientEmails: string[] }> {
-    const to = await profileService.getEmailsByRoles(input.recipientRoles);
-
-    if (to.length === 0) {
+  async sendCampaignEmail(
+    subject: string,
+    body: string,
+    recipientEmails: string[],
+  ): Promise<{ id?: string; recipientEmails: string[] }> {
+    if (recipientEmails.length === 0) {
       throw new ApiError(
         "No recipients found for the selected audiences",
         400,
@@ -155,22 +152,22 @@ class EmailService {
 
     const baseHtml = await render(
       createElement(CampaignEmailTemplate, {
-        subject: input.subject,
-        body: input.body,
+        subject,
+        body,
       }),
     );
     const html = appendMarketingUnsubscribeFooter(baseHtml);
 
-    await this.ensureRecipientsInSegment(to, this.campaignSegmentId);
+    await this.ensureRecipientsInSegment(recipientEmails, this.campaignSegmentId);
     try {
       const { id } = await this.sendMarketingBroadcast({
         segmentId: this.campaignSegmentId,
-        subject: input.subject,
+        subject,
         html,
       });
-      return { id, recipientEmails: to };
+      return { id, recipientEmails };
     } catch (err) {
-      await this.removeRecipientsFromSegment(to);
+      await this.removeRecipientsFromSegment(recipientEmails);
       throw err;
     }
   }
